+++ packages/react-mui-shared/src/components/PangeaFields/FieldsForm/types/index.ts
@@ -140,6 +140,7 @@ export interface MultilineFieldSchemaProps {
 
 export interface DateTimeFieldSchemaProps {
   type: "dateTime";
+  maxDate?: any;
 }
 
 export interface SwitchFieldSchemaProps {