+++ packages/react-mui-store-file-viewer/src/stories/StoreFileViewer.stories.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import dayjs from "dayjs";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
 import StoreFileViewer, {